LLM Course/7.1.Token classification/requirements.txt
# Token Classification框架依赖包
# 核心深度学习框架
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# 数据处理和科学计算
numpy>=1.24.0
pandas>=2.0.0

# 评估指标
evaluate>=0.4.0
scikit-learn>=1.3.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置文件处理
PyYAML>=6.0

# 进度条
tqdm>=4.65.0

# 日志
colorlog>=6.7.0

# 可选：加速训练
# accelerate>=0.20.0  # 如需分布式训练
# tensorboard>=2.13.0  # 如需TensorBoard日志

# 可选：模型量化和优化
# optimum>=1.8.0
# bitsandbytes>=0.40.0  # 如需量化
